Erc2771 (#49)
* Use EnumerableSet for operator approvals

This will allow us to get all the approvals for a user from a frontend dashboard. We will have to get all the user addresses and then call this function to get all of their approvals.

* Add owner mapping for validations and delegations (#37)

* add owner mappings for validators and delegators

Add mappings from owner to validations and owner to delegations.

These are created when `initiate` is called and removed when the last rewards are distributed, or when complete validator removal is called and there are no rewards left

* delete comments

* Consistent function namees

* reorder functions and limit function visibility

* more function reordering

* test delegation/validations are properly removed

* Add test for getRewardsMintedForEpoch

* WIP

* ERC2771 Gasless

* Protect setTrustedForwarder

* fix-duplicate-tests

* fix: don't delete validations on complete step

* Add back context_init for clarity

---------

Co-authored-by: Johnny Gault <john@multisiglabs.org>